extract common functions to helpers_test.go
